back/requirements.txt
asgiref==3.3.1
Django==3.2.3
django-cors-headers==3.2.1
django-crispy-forms==1.11.1
django-extensions==3.1.2
django-filter==2.4.0
django-filters==0.2.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
generics==3.3.0
Pillow==7.1.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pydot==1.4.2
pydotplus==2.0.2
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
sqlparse==0.4.1
